// icd-annotation-frontend/src/localComponents/types/enums.d.ts
export enum Sex {
    Both = "Both",
    Male = "Male",
    Female = "Female",
    Other = "Other"
}
// ...
export enum TimeOfYear {
    Spring = "Spring",
    Summer = "Summer",
    Autumn = "Autumn",
    Winter = "Winter"
}